QPM_files
The website can generate the question papers in handy of time will be
greatly useful for the academic institutions<br/> which requires to take
the objective type examination regularly and finds it tough to create a
properly formatted paper<br/> with the appropriate set of question.

Teachers can make account,manage the questions in a easy Way !!!

Categorization in classes,departments and subjects make it effecient to
manage and maintain.